Make utils static
